M4: Split AdvancedFilterClientInput mega-test - Extract onBlur callback
Reduced main test file from 510 lines to 372 lines (138 line reduction).
Created on-blur-callback.unit.test.tsx with 188 lines (4 tests).

Extracted test group:
- onBlur callback (4 tests covering IDIR/BCEIDBUSINESS blur behavior)

All 1421 tests passing. Test file count now 123.
